create working dashboard + components
