+++ package.json
@@ -25,6 +25,7 @@
     "react-hook-form": "^7.56.1",
     "react-icons": "^5.5.0",
     "tailwind": "^4.0.0",
+    "tailwind-merge": "^3.3.1",
     "zod": "^3.24.3"
   },
   "devDependencies": {
