+++ tsconfig.json
@@ -23,6 +23,6 @@
       "@/*": ["src/*", "./public/*"]
     }
   },
-  "include": ["next-env.d.ts", ".next/types/**/*.ts", "**/*.ts", "**/*.tsx"],
+  "include": ["next-env.d.ts", ".next/types/**/*.ts", "**/*.ts", "**/*.tsx", ".next/types/app/careprovider-dashboard/page.tsx", ".next/types/app/layout.tsx"],
   "exclude": ["node_modules"]
 }
